pyproject.toml
[build-system]
requires = ["setuptools>=75.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "my-agent"
version = "0.1.0"
description = "AI Agent framework — multi-agent collaboration + self-evolution"
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}
dependencies = [
    "httpx>=0.28.0",
    "anthropic>=0.40.0",
    "pyyaml>=6.0",
    "jieba>=0.42",
    "prompt_toolkit>=3.0",
]

[project.optional-dependencies]
web = [
    "playwright>=1.50.0",
]
test = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
